Use jsons 1.4.2
ramonhagenaars/jsons#133 was resolved super quick, so we can use that
now.

The hashed keys don't look great, but they're not user-hostile (the
hash can be *anything* when deserialising as long as the value is
still an integer array of 2--3 values). Besides, this removes a tonne
of boilerplate when modifying the config classes.

Thanks so much @ramonhagenaars! (amazing package)